fix(investigation-bias): Fix event.type in the investigation bias and move around some functions
