+++ src/sentry/dynamic_sampling/rules/biases/custom_rule_bias.py
@@ -31,4 +31,5 @@ class CustomRuleBias(Bias):
                     },
                 }
             )
+
         return ret_val
